requirements.txt
bleach==6.2.0
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
pydantic==2.11.3
alembic==1.15.2
sqlalchemy==2.0.40
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
email-validator==2.2.0
passlib==1.7.4
bcrypt==4.3.0
marshmallow==4.0.0
webargs==8.7.0
flask-pydantic==0.13.1
fastapi==0.115.12
sqlalchemy-stubs==0.4
flask-swagger-ui==4.11.1
werkzeug==3.1.3
PyJWT==2.10.1
python-jose==3.4.0
requests==2.32.3